chore: Initial spike of i18n discovery
This commit adds an example plugin that has most of it's UI elements set
up to use the translation engine exposed by the Backstage core API.  The
main app is also updated to enable translations for the navigation.
These translations are set up in the main app's initialization function,
where override messages are brought in for both English and German.
This change also removes the ADR plugin that was previously set up
during some initial tire-kicking of the implementation.